Added fee structs to common structs module
